% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subarea.R
\name{extract_subarea}
\alias{extract_subarea}
\title{FUNCTION_TITLE}
\usage{
extract_subarea(in_mosaics_folder, in_mask, subset_name, out_folder)
}
\arguments{
\item{in_mosaics_folder}{PARAM_DESCRIPTION}

\item{in_mask}{PARAM_DESCRIPTION}

\item{subset_name}{PARAM_DESCRIPTION}

\item{out_folder}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 mosaic_folder <- "/home/lb/my_data/prasia/mosaics/ordered"
 out_folder  <- "/home/lb/my_data/prasia/mosaics/ordered/subsets/"
 subset_name <- "Nueva_Ecija"
 in_country  <- "PHL"
 boundmask   <- sprawl::get_boundaries(in_country, level = 1) \%>\%
 sf::st_as_sf() \%>\%
 dplyr::filter(NAME_1 == "Nueva Ecija") \%>\%
 sf::st_combine()
 extract_subarea(mosaic_folder,
                boundmask,
                subset_name,
                out_folder)
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
