% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenorice_GUI.R
\name{phenorice_GUI}
\alias{phenorice_GUI}
\title{phenorice_GUI}
\usage{
phenorice_GUI()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
