% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_create_ordered_tiffs.R
\name{pr_create_ordered_tiffs}
\alias{pr_create_ordered_tiffs}
\title{pr_create_ordered_tiffs}
\usage{
pr_create_ordered_tiffs(mosaics_folder, patterns, out_folder)
}
\arguments{
\item{mosaics_folder}{PARAM_DESCRIPTION}

\item{patterns}{PARAM_DESCRIPTION}

\item{out_folder}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 # # Folder containing the mosaics by year and parameter (e.g., sos_1_2016, sos_2_2016, etc)
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
