% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_create_vrt_mosaics.R
\name{pr_create_vrt_mosaics}
\alias{pr_create_vrt_mosaics}
\title{pr_create_vrt_mosaics}
\usage{
pr_create_vrt_mosaics(mosaics_folder, pattern)
}
\arguments{
\item{mosaics_folder}{PARAM_DESCRIPTION}

\item{pattern}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper to create virtual mosaics used in create_ordered_tiffs
}
\details{
DETAILS
}
\examples{
\dontrun{
 #EXAMPLE1
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
