% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_extract.R
\name{pr_extract}
\alias{pr_extract}
\title{pr_extract}
\usage{
pr_extract(main_folder, region)
}
\arguments{
\item{main_folder}{location of the "Data" folder (e.g., "/home/lb/my_data/prasia/Data")}

\item{region}{region to extract(e.g., "Region_3_-\emph{Central_Luzon").
If "All", all regions are extracted. Multoiple values
are possible. (e.g., region = c("Region_3}-_Central_Luzon", "Chippalonga"))}
}
\value{
\itemize{
\item Cropped rasters are stored in the \code{Data/subsets/_region_name_/orig} and
\code{Data/subsets/_region_name_/decirc} folders. #'
\item Extracted data are saved in the \code{Data/subsets/_region_name_/RData} folder. One
separate RData file is created for each PhenoRice parameter (e.g., "Dhaka_sos_stats.RData",
"Dhaka_eos_stats.RData", "Dhaka_sos_decirc_stats.RData")
}
}
\description{
Extract data from PhenoRice mosaics
}
\examples{
\dontrun{
# set the main folder ----
main_folder <- "/home/lb/my_data/prasia/Data"

# Suppose you want to extract data for Region: "Region_3_-_Central_Luzon" :
# --> extract it from the full shapefile
Region_name <- "Region_3_-_Central_Luzon"

# "1) run "pr_extract" on the region ----

pr_extract(main_folder,
          region = "Region_name")
}

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
