% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_extract_idl_outs.R
\name{pr_extract_idl_outs}
\alias{pr_extract_idl_outs}
\title{FUNCTION_TITLE}
\usage{
pr_extract_idl_outs(in_folder, out_folder, start_year, end_year,
  overwrite = FALSE)
}
\arguments{
\item{in_folder}{PARAM_DESCRIPTION}

\item{out_folder}{PARAM_DESCRIPTION}

\item{start_year}{PARAM_DESCRIPTION}

\item{end_year}{PARAM_DESCRIPTION}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
