% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_mutate_when.R
\name{pr_mutate_when}
\alias{pr_mutate_when}
\title{pr_mutate_when}
\usage{
pr_mutate_when(data, ...)
}
\arguments{
\item{data}{a \code{data.frame} (or \code{tibble} or \code{data.table})}

\item{...}{expression to be evaluated to filter on rows}
}
\value{
\code{data frame}
}
\description{
mutate a column within a dplyr pype only on a subset of rows
specified by a condition
Retrieved from a StackOverflow answer (https://stackoverflow.com/a/34170176/6871135)
}
\details{
DETAILS
}
\examples{
\dontrun{
 library(magrittr)
 mtcars \%>\% mutate_when(
            mpg > 22,    list(cyl = 100),
            disp == 160, list(cyl = 200))

}
}
\author{
Kevin Ushey
}
