% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_riceatlas.R
\name{read_riceatlas}
\alias{read_riceatlas}
\title{FUNCTION_TITLE}
\usage{
read_riceatlas(in_txt, in_riceatlas_shp, ISO)
}
\arguments{
\item{in_txt}{PARAM_DESCRIPTION}

\item{in_riceatlas_shp}{PARAM_DESCRIPTION}

\item{ISO}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
