% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_year_mosaics.R
\name{rename_year_mosaics}
\alias{rename_year_mosaics}
\title{FUNCTION_TITLE}
\usage{
rename_year_mosaics(mosaics_folder, out_folder)
}
\arguments{
\item{mosaics_folder}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
